Move files around to better align with their purpose
