+++ src/components/clusters/wizards/rosa/LogForwarding/logForwardingNaming.ts
